refactor: rename primitives
