+++ packages/pure-parse/src/guard/json.ts
@@ -1,5 +1,5 @@
 import { arrayGuard, partialRecord, union } from './validation'
-import { isBoolean, isNull, isNumber, isString } from './guards'
+import { isBoolean, isNull, isNumber, isString } from './primitives'
 import { JsonValue } from '../common'
 
 export const isJsonValue = (data: unknown): data is JsonValue =>
